Inline `-preset.js` file
There is only one transform in this codemod so all this complexity seems unnecessary
